requirements.txt
fastapi
uvicorn
qdrant-client
langchain-google-vertexai
python-dotenv
pypdf
